requirements.txt
platformio==6.1.15
pyserial==3.5
prompt_toolkit==3.0.43
scapy==2.5.0